feat: create reminder store to persist reminders between renders
