+++ screens/AppStack/CreateReminderScreen.tsx
@@ -9,7 +9,7 @@ import {
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 import * as yup from 'yup'
-import { useCallback, useState, useEffect, useRef } from 'react'
+import { useCallback, useState, useEffect } from 'react'
 import DateTimePicker from '@react-native-community/datetimepicker'
 
 import {
